requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
flask==3.0.3
importlib-metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
flask_cors==3.0.7
kafka-python
MarkupSafe==2.1.5
werkzeug==3.0.2
zipp==3.18.1
gunicorn
gevent
flask-socketio
eventlet==0.30.1
psycopg2